+++ packages/components/src/Icon/index.ts
@@ -25,3 +25,4 @@
  */
 
 export * from './Icon'
+export * from './IconPlaceholder'
