+++ packages/components/src/Menu/MenuItem.tsx
@@ -30,10 +30,10 @@ import isFunction from 'lodash/isFunction'
 import styled from 'styled-components'
 import React, { FC, ReactNode, useContext, useState, useEffect } from 'react'
 import { Placement } from '@popperjs/core'
-import { IconPlaceholder, ListItemDetail } from '../List'
+import { ListItemDetail } from '../List'
 import { Paragraph } from '../Text'
 import { useID } from '../utils/useID'
-import { Icon } from '../Icon'
+import { Icon, IconPlaceholder } from '../Icon'
 import { Tooltip } from '../Tooltip'
 import { MenuContext, MenuItemContext } from './MenuContext'
 import { MenuItemLayout } from './MenuItemLayout'
